+++ covalent/executor/executor_plugins/local.py
@@ -27,10 +27,12 @@ This is a plugin executor module; it is loaded if found and properly structured.
 
 import os
 from concurrent.futures import ProcessPoolExecutor
+from pathlib import Path
 from typing import Any, Callable, Dict, List
 
 # Relative imports are not allowed in executor plugins
 from covalent._shared_files import TaskCancelledError, TaskRuntimeError, logger
+from covalent._shared_files.config import get_config, set_config
 from covalent.executor import BaseExecutor
 
 # Store the wrapper function in an external module to avoid module
@@ -49,6 +51,7 @@ _EXECUTOR_PLUGIN_DEFAULTS = {
     "cache_dir": os.path.join(
         os.environ.get("XDG_CACHE_HOME") or os.path.join(os.environ["HOME"], ".cache"), "covalent"
     ),
+    "workdir": os.path.join(os.environ["HOME"], "covalent", "workdir"),
 }
 
 proc_pool = ProcessPoolExecutor()
@@ -58,6 +61,18 @@ class LocalExecutor(BaseExecutor):
     """
     Local executor class that directly invokes the input function.
     """
+
+    def __init__(self, workdir: str = "", *args, **kwargs) -> None:
+        if not workdir:
+            try:
+                workdir = get_config("executors.local.workdir")
+            except KeyError:
+                workdir = _EXECUTOR_PLUGIN_DEFAULTS["workdir"]
+                set_config("executors.local.workdir", workdir)
+
+        super().__init__(*args, **kwargs)
+        self.workdir = workdir
+        Path(self.workdir).mkdir(parents=True, exist_ok=True)
 
     def run(self, function: Callable, args: List, kwargs: Dict, task_metadata: Dict) -> Any:
         """
@@ -81,7 +96,7 @@ class LocalExecutor(BaseExecutor):
             raise TaskCancelledError
 
         # Run the target function in a separate process
-        fut = proc_pool.submit(io_wrapper, function, args, kwargs)
+        fut = proc_pool.submit(io_wrapper, self.workdir, function, args, kwargs)
 
         output, worker_stdout, worker_stderr, tb = fut.result()