+++ covalent/executor/utils/wrappers.py
@@ -23,6 +23,7 @@ Helper functions for the local executor
 """
 
 import io
+import os
 import traceback
 from contextlib import redirect_stderr, redirect_stdout
 from enum import Enum
@@ -39,14 +40,18 @@ class Signals(Enum):
     EXIT = 2
 
 
-def io_wrapper(fn: Callable, args: List, kwargs: Dict) -> Tuple[Any, str, str, str]:
+def io_wrapper(workdir: str, fn: Callable, args: List, kwargs: Dict) -> Tuple[Any, str, str, str]:
     """Wrapper function to execute the given function in a separate
     process and capture stdout and stderr"""
     with redirect_stdout(io.StringIO()) as stdout, redirect_stderr(io.StringIO()) as stderr:
         try:
+            current_dir = os.getcwd()
+            os.chdir(workdir)
             output = fn(*args, **kwargs)
             tb = ""
         except Exception as ex:
             output = None
             tb = "".join(traceback.TracebackException.from_exception(ex).format())
+        finally:
+            os.chdir(current_dir)
     return output, stdout.getvalue(), stderr.getvalue(), tb