+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Added the `CloudResourceManager` class
+- Ability to specify a `workdir` for `local` and `dask` executors
 
 ### Tests